requirements.txt
asgiref==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
idna==3.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
platformdirs==3.5.1
protobuf==3.20.3
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
redis==4.5.5